Moved all Panels to separate classes
