+++ src/OSATE AADL AGREE Extension/src/org/osate/pluginsample/ui/HeaderPanel.java
@@ -1,0 +1,19 @@
+package org.osate.pluginsample.ui;
+
+import java.awt.Font;
+
+import javax.swing.JLabel;
+import javax.swing.JPanel;
+
+public class HeaderPanel extends JPanel {
+	private JLabel headerLabel = new JLabel();
+	
+	public HeaderPanel() {
+		headerLabel.setFont(new java.awt.Font("Arial", Font.PLAIN, 32));
+		add(headerLabel);
+	}
+	
+	public void setHeaderLabel(String newLabel) {
+		headerLabel.setText(newLabel.substring(0, 1).toUpperCase() + newLabel.substring(1));
+	}
+}
